# server/openapi/api-validaciones.yaml
openapi: 3.1.1
info:
  title: API de validaciones del lector de cédula CO
  version: 1.0.0
  summary: Crear, alimentar, consultar y suprimir validaciones de la cédula colombiana.
  description: |
    API del servidor de respaldo del lector de cédula colombiana (amarilla con PDF417 y digital
    con MRZ). Este documento es la fuente de verdad del contrato: la aplicación lo sirve tal cual en
    `GET /openapi.json`.

    Flujo: `POST /v1/validations` crea una validación en `pending` y devuelve `upload.url`; el
    cliente envía anverso, reverso y, solo con comparación facial, selfie en una única petición
    `multipart/form-data` a esa URL; la respuesta trae la validación en estado terminal
    (`success`, `failure` o `review`). Si se configuró `webhook_url`, el servidor notifica el
    evento `validation.completed` firmado con HMAC-SHA256.

    Privacidad (Ley 1581 de 2012): toda creación exige la autorización expresa del titular
    (`autorizacion.datos`) y, para comparación facial, la de datos sensibles
    (`autorizacion.sensibles`). Las imágenes se procesan solo en memoria y nunca se guardan; ninguna
    respuesta ni webhook contiene imágenes ni biometría. El resultado se elimina con `DELETE` o al
    vencer la retención (`expires_at`).

    Sandbox: las claves `sk_test_` operan en modo sandbox con resultados sintéticos deterministas
    (`sandbox_scenario`); las claves `sk_live_` operan en modo live.

...
  contact:
    name: Equipo del lector de cédula CO
...
  license:
...
servers:
...
security:
  - claveApi: []
paths:
...
  /v1/validations:
    post:
      operationId: crearValidacion
      tags: [validaciones]
      summary: Crear una validación
      description: |
        Crea una validación en estado `pending` y devuelve la URL firmada de subida. Exige la
        autorización expresa del titular. El cuerpo admite como máximo 16 384 bytes y rechaza campos
        no declarados. Con `Idempotency-Key`, repetir la misma petición devuelve la misma respuesta
        durante 86 400 s.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrearValidacion"
            examples:
              minima:
                summary: Creación mínima en sandbox
                value:
                  document_type: co_national-id-2000
                  autorizacion:
                    datos: true
                    sensibles: false
                    version_texto: "2026-10-01"
                    otorgada_en: "2026-10-06T15:19:00Z"
      responses:
        "201":
          description: Validación creada en `pending`.
          headers:
            Location:
              $ref: "#/components/headers/Location"
            Idempotent-Replayed:
              $ref: "#/components/headers/IdempotentReplayed"
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            Cache-Control:
              $ref: "#/components/headers/CacheControl"
            X-Content-Type-Options:
              $ref: "#/components/headers/XContentTypeOptions"
            Referrer-Policy:
              $ref: "#/components/headers/ReferrerPolicy"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Validation"
              examples:
                pendiente:
                  summary: Validación recién creada
                  value:
                    id: val_0123456789abcdef0123456789abcdef
                    object: validation
                    sandbox: true
                    document_type: co_national-id-2000
                    face_match: false
                    status: pending
                    declined_reason: null
                    checks: []
                    document: null
                    autorizacion:
                      datos: true
                      sensibles: false
                      version_texto: "2026-10-01"
                      otorgada_en: "2026-10-06T15:19:00Z"
                      registrada_en: "2026-10-06T15:20:00Z"
                    upload:
                      url: "https://api.lector-cedula.example/v1/validations/val_0123456789abcdef0123456789abcdef/images?token=AAAAAGkm0yxzaW50ZXRpY28tc29sby1lamVtcGxvLW5vLXZhbGlkbw"
                      expires_at: "2026-10-06T15:35:00Z"
                    webhook_url: null
                    return_url: "https://app-a.example/volver"
                    hosted_url: "https://api.lector-cedula.example/v/AAAAAGqaElwBI0VniavN7wEjRWeJq83vc2ludGV0aWNvLXNvbG8tZWplbXBsby1uby12YWxpZG8"
                    created_at: "2026-10-06T15:20:00Z"
                    updated_at: "2026-10-06T15:20:00Z"
                    completed_at: null
                    expires_at: null
          links:
            ConsultarValidacion:
              operationId: obtenerValidacion
              description: Consulta la validación creada.
              parameters:
                id: "$response.body#/id"
            SubirImagenes:
              operationId: subirImagenes
              description: Sube las imágenes de la validación creada.
              parameters:
                id: "$response.body#/id"
            SuprimirValidacion:
              operationId: suprimirValidacion
              description: Suprime la validación creada.
              parameters:
                id: "$response.body#/id"
        "400":
          $ref: "#/components/responses/SolicitudInvalida"
        "401":
          $ref: "#/components/responses/NoAutorizado"
        "403":
          $ref: "#/components/responses/Prohibido"
        "409":
          $ref: "#/components/responses/Conflicto"
        "413":
          $ref: "#/components/responses/DemasiadoGrande"
        "422":
          $ref: "#/components/responses/NoProcesable"
        "429":
          $ref: "#/components/responses/DemasiadasPeticiones"
        "500":
          $ref: "#/components/responses/ErrorInterno"
  /v1/validations/{id}:
    parameters:
      - $ref: "#/components/parameters/IdValidacion"
    get:
      operationId: obtenerValidacion
      tags: [validaciones]
      summary: Consultar una validación
      description: |
        Devuelve el estado actual de la validación. Un identificador inexistente, mal formado,
        suprimido, vencido o de otro cliente responde 404 con el mismo cuerpo salvo `request_id`.
      responses:
        "200":
          description: Estado actual de la validación.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            Cache-Control:
              $ref: "#/components/headers/CacheControl"
            X-Content-Type-Options:
              $ref: "#/components/headers/XContentTypeOptions"
            Referrer-Policy:
              $ref: "#/components/headers/ReferrerPolicy"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Validation"
              examples:
                exitosa:
                  summary: Cédula amarilla validada en sandbox
                  value:
                    id: val_0123456789abcdef0123456789abcdef
                    object: validation
                    sandbox: true
                    document_type: co_national-id-2000
                    face_match: false
                    status: success
                    declined_reason: null
                    checks:
                      - category: image_quality
                        status: passed
                        reasons: []
                      - category: data_validation
                        status: passed
                        reasons: []
                      - category: data_consistency
                        status: passed
                        reasons: []
                      - category: document_liveness
                        status: passed
                        reasons: []
                    document:
                      type: co_national-id-2000
                      document_number: "9999123456"
                      first_surname: PEÑA
                      second_surname: DE LA OSSA
                      first_name: FICTICIA
                      second_name: null
                      sex: F
                      date_of_birth: "1990-02-28"
                      place_of_birth:
                        divipol_department: "01"
                        divipol_municipality: "001"
                      blood_type: AB-
                      date_of_issue: null
                      place_of_issue: null
                      date_of_expiry: null
                      sources: [pdf417]
                      warnings: []
                    autorizacion:
                      datos: true
                      sensibles: false
                      version_texto: "2026-10-01"
                      otorgada_en: "2026-10-06T15:19:00Z"
                      registrada_en: "2026-10-06T15:20:00Z"
                    upload: null
                    webhook_url: null
                    return_url: null
                    hosted_url: null
                    created_at: "2026-10-06T15:20:00Z"
                    updated_at: "2026-10-06T15:20:00Z"
                    completed_at: "2026-10-06T15:20:00Z"
                    expires_at: "2026-10-07T15:20:00Z"
        "401":
          $ref: "#/components/responses/NoAutorizado"
        "403":
          $ref: "#/components/responses/Prohibido"
        "404":
          $ref: "#/components/responses/NoEncontrado"
        "429":
          $ref: "#/components/responses/DemasiadasPeticiones"
        "500":
          $ref: "#/components/responses/ErrorInterno"
    delete:
      operationId: suprimirValidacion
      tags: [validaciones]
      summary: Suprimir una validación
      description: |
        Elimina la validación y su resultado por revocación de la autorización del titular y
        cancela los webhooks pendientes. Después, toda operación sobre el mismo `id` responde 404.
      responses:
        "204":
          description: Validación suprimida. Sin cuerpo.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            Cache-Control:
              $ref: "#/components/headers/CacheControl"
            X-Content-Type-Options:
              $ref: "#/components/headers/XContentTypeOptions"
            Referrer-Policy:
              $ref: "#/components/headers/ReferrerPolicy"
        "401":
          $ref: "#/components/responses/NoAutorizado"
        "403":
          $ref: "#/components/responses/Prohibido"
        "404":
          $ref: "#/components/responses/NoEncontrado"
        "429":
          $ref: "#/components/responses/DemasiadasPeticiones"
        "500":
          $ref: "#/components/responses/ErrorInterno"
  /v1/validations/{id}/images:
    parameters:
      - $ref: "#/components/parameters/IdValidacion"
    post:
      operationId: subirImagenes
      tags: [validaciones]
      summary: Subir las imágenes y procesar la validación
      description: |
        Recibe en una sola petición `multipart/form-data` las partes `front` y `back` y, solo si la
        validación pidió comparación facial, `selfie`. Cualquier otra parte o campo de texto (por ejemplo
        `document` o `nuip`) se rechaza con 422 (`unexpected_field` o `unexpected_part`): el resultado
        lo calcula solo el servidor a partir de las imágenes. Cada parte es JPEG o PNG de 1 a 8 388 608
        bytes; el cuerpo completo mide como máximo 20 971 520 bytes. La validación se procesa de
        forma síncrona y en memoria: ninguna imagen sobrevive a la petición. Se autentica con el
        token de `upload.url` (vigencia de 900 s, un solo uso) o con la clave que creó la
        validación.
      security:
        - tokenSubida: []
        - claveApi: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ImagenesValidacion"
            encoding:
              front:
                contentType: image/jpeg, image/png
              back:
                contentType: image/jpeg, image/png
              selfie:
                contentType: image/jpeg, image/png
      responses:
        "200":
          description: Validación procesada, en estado terminal.
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
            Cache-Control:
              $ref: "#/components/headers/CacheControl"
            X-Content-Type-Options:
              $ref: "#/components/headers/XContentTypeOptions"
            Referrer-Policy:
              $ref: "#/components/headers/ReferrerPolicy"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Validation"
        "401":
          $ref: "#/components/responses/NoAutorizado"
        "403":
          $ref: "#/components/responses/Prohibido"
        "404":
          $ref: "#/components/responses/NoEncontrado"
        "409":
          $ref: "#/components/responses/Conflicto"
        "413":
          $ref: "#/components/responses/DemasiadoGrande"
        "415":
          $ref: "#/components/responses/TipoNoAdmitido"
        "422":
          $ref: "#/components/responses/NoProcesable"
        "429":
          $ref: "#/components/responses/DemasiadasPeticiones"
        "500":
          $ref: "#/components/responses/ErrorInterno"
        "503":
          $ref: "#/components/responses/MotorNoDisponible"
...
components:
  securitySchemes:
    claveApi:
      type: http
      scheme: bearer
      description: |
        Clave de API en `Authorization: Bearer <clave>`. El prefijo `sk_test_` selecciona el modo
        sandbox y `sk_live_` el modo live. Cada validación solo es visible para la clave que la creó.
    tokenSubida:
      type: apiKey
      in: query
      name: token
      description: |
        Token firmado de `upload.url`. Autoriza solo la subida a esa validación, vence en 900 s y
        sirve una sola vez.
  parameters:
    IdValidacion:
      name: id
      in: path
      required: true
      description: Identificador de la validación.
      schema:
        type: string
        pattern: "^val_[0-9a-f]{32}$"
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Clave de idempotencia de 1 a 255 caracteres ASCII imprimibles, con alcance por clave de API
        y vigencia de 86 400 s. Repetir clave y cuerpo devuelve la misma respuesta con
        `Idempotent-Replayed: true`; repetir la clave con otro cuerpo responde 422.
      schema:
        type: string
        minLength: 1
        maxLength: 255
        pattern: "^[\\x20-\\x7E]+$"
  headers:
    XRequestId:
      description: Identificador de la petición (UUID versión 4); coincide con `request_id` en los errores.
      required: true
      schema:
        type: string
        format: uuid
    CacheControl:
      description: Las respuestas nunca se guardan en cachés.
      required: true
      schema:
        type: string
        const: no-store
    XContentTypeOptions:
      description: Impide que el navegador adivine el tipo de contenido.
      required: true
      schema:
        type: string
        const: nosniff
    ReferrerPolicy:
      description: Impide que la URL con el token de subida viaje en `Referer`.
      required: true
      schema:
        type: string
        const: no-referrer
    Location:
      description: Ruta de la validación creada, `/v1/validations/{id}`.
      required: true
      schema:
        type: string
        pattern: "^/v1/validations/val_[0-9a-f]{32}$"
    IdempotentReplayed:
      description: Presente con `true` cuando la respuesta es la repetición de una creación anterior.
      required: false
      schema:
        type: string
        const: "true"
...
    WwwAuthenticate:
      description: Esquema de autenticación requerido.
      required: true
      schema:
        type: string
        const: Bearer
  responses:
    SolicitudInvalida:
      description: Cabecera inválida (por ejemplo, `invalid-idempotency-key`).
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Cache-Control:
          $ref: "#/components/headers/CacheControl"
        X-Content-Type-Options:
          $ref: "#/components/headers/XContentTypeOptions"
        Referrer-Policy:
          $ref: "#/components/headers/ReferrerPolicy"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NoAutorizado:
      description: Falta la clave de API o no está configurada (`unauthorized`).
      headers:
        WWW-Authenticate:
          $ref: "#/components/headers/WwwAuthenticate"
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Cache-Control:
          $ref: "#/components/headers/CacheControl"
        X-Content-Type-Options:
          $ref: "#/components/headers/XContentTypeOptions"
        Referrer-Policy:
          $ref: "#/components/headers/ReferrerPolicy"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Prohibido:
      description: |
        Token de subida inválido o vencido (`upload-token-invalid`, `upload-token-expired`), o petición
        de navegador cuyo `Origin` no está en los orígenes configurados para la clave
        (`origin-not-allowed`). Las peticiones de servidor no envían `Origin`.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Cache-Control:
          $ref: "#/components/headers/CacheControl"
        X-Content-Type-Options:
          $ref: "#/components/headers/XContentTypeOptions"
        Referrer-Policy:
          $ref: "#/components/headers/ReferrerPolicy"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NoEncontrado:
      description: |
        La validación no existe, está mal identificada, fue suprimida, venció o es de otro cliente
        (`not-found`). El cuerpo es el mismo en todos los casos salvo `request_id`.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Cache-Control:
          $ref: "#/components/headers/CacheControl"
        X-Content-Type-Options:
          $ref: "#/components/headers/XContentTypeOptions"
        Referrer-Policy:
          $ref: "#/components/headers/ReferrerPolicy"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          examples:
            noEncontrada:
              summary: Validación inexistente
              value:
                type: https://lector-cedula.example/problemas/not-found
                title: Recurso no encontrado
                status: 404
                detail: El recurso solicitado no existe o no está disponible para esta clave.
                code: not-found
                request_id: 3f2c9a7e-5b1d-4c8e-9f0a-2b6d4e8c1a3f
    Conflicto:
      description: |
        La validación no está en `pending` (`validation-not-pending`) o hay una creación en curso
        con la misma clave de idempotencia (`idempotency-key-in-progress`).
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Cache-Control:
          $ref: "#/components/headers/CacheControl"
        X-Content-Type-Options:
          $ref: "#/components/headers/XContentTypeOptions"
        Referrer-Policy:
          $ref: "#/components/headers/ReferrerPolicy"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    DemasiadoGrande:
      description: El cuerpo o una imagen supera el límite (`request-too-large`, `image-too-large`).
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Cache-Control:
          $ref: "#/components/headers/CacheControl"
        X-Content-Type-Options:
          $ref: "#/components/headers/XContentTypeOptions"
        Referrer-Policy:
          $ref: "#/components/headers/ReferrerPolicy"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    TipoNoAdmitido:
      description: La imagen no es JPEG ni PNG o su firma no coincide con el tipo declarado (`unsupported-image-type`).
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Cache-Control:
          $ref: "#/components/headers/CacheControl"
        X-Content-Type-Options:
          $ref: "#/components/headers/XContentTypeOptions"
        Referrer-Policy:
          $ref: "#/components/headers/ReferrerPolicy"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NoProcesable:
      description: |
        El cuerpo no cumple el contrato (`invalid-request`) o reutiliza una clave de idempotencia con
        otro cuerpo (`idempotency-key-reused`). `errors` señala cada campo con un JSON Pointer.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Cache-Control:
          $ref: "#/components/headers/CacheControl"
        X-Content-Type-Options:
          $ref: "#/components/headers/XContentTypeOptions"
        Referrer-Policy:
          $ref: "#/components/headers/ReferrerPolicy"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
          examples:
            camposFaltantes:
              summary: Cuerpo vacío
              value:
                type: https://lector-cedula.example/problemas/invalid-request
                title: Solicitud inválida
                status: 422
                detail: El cuerpo no cumple el contrato. Revise `errors`.
                code: invalid-request
                request_id: 3f2c9a7e-5b1d-4c8e-9f0a-2b6d4e8c1a3f
                errors:
                  - pointer: /document_type
                    code: required
                  - pointer: /autorizacion
                    code: required
    DemasiadasPeticiones:
      description: Se superó el límite de peticiones por minuto de la clave (`rate-limited`).
      headers:
        Retry-After:
          $ref: "#/components/headers/RetryAfter"
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Cache-Control:
          $ref: "#/components/headers/CacheControl"
        X-Content-Type-Options:
          $ref: "#/components/headers/XContentTypeOptions"
        Referrer-Policy:
          $ref: "#/components/headers/ReferrerPolicy"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ErrorInterno:
      description: Error interno (`internal-error`). Nunca incluye trazas ni valores del cliente.
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Cache-Control:
          $ref: "#/components/headers/CacheControl"
        X-Content-Type-Options:
          $ref: "#/components/headers/XContentTypeOptions"
        Referrer-Policy:
          $ref: "#/components/headers/ReferrerPolicy"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    MotorNoDisponible:
      description: El modo live no tiene motor de procesamiento configurado (`engine-unavailable`).
      headers:
        X-Request-Id:
          $ref: "#/components/headers/XRequestId"
        Cache-Control:
          $ref: "#/components/headers/CacheControl"
        X-Content-Type-Options:
          $ref: "#/components/headers/XContentTypeOptions"
        Referrer-Policy:
          $ref: "#/components/headers/ReferrerPolicy"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    DocumentType:
      type: string
      description: Tipo de documento. `co_national-id-2000` es la cédula amarilla con PDF417; `co_national-id-2020`, la digital con MRZ.
      enum:
        - co_national-id-2000
        - co_national-id-2020
...
    CrearValidacion:
      type: object
      description: Cuerpo de la creación (como máximo 16 384 bytes; no admite campos no declarados).
      required: [document_type, autorizacion]
      additionalProperties: false
      properties:
        document_type:
          $ref: "#/components/schemas/DocumentType"
        autorizacion:
          $ref: "#/components/schemas/Autorizacion"
        face_match:
          type: boolean
          default: false
          description: Pide comparación facial con selfie. Exige `autorizacion.sensibles` en `true`.
        webhook_url:
          type: string
          format: uri
          maxLength: 2048
          pattern: "^https://"
          description: |
            URL `https` que recibe el evento `validation.completed`. Sin credenciales y con un host
            que no sea una IP literal; si al entregar resuelve a una red interna, el envío se bloquea.
        sandbox_scenario:
          $ref: "#/components/schemas/SandboxScenario"
        return_url:
          type: string
          maxLength: 2048
          pattern: "^[a-z][a-z0-9+.-]*://"
          description: |
            Destino al terminar la sesión alojada (`hosted_url`). Debe ser exactamente una de las URL
            `retornos` configuradas para la clave (`https` o deeplink con esquema propio); si no, 422
            con `return_url_not_allowed`. Solo `validation_id` y `estado` viajan al retornar.
...
    Validation:
      type: object
      description: |
        Validación de un documento. En `pending`, `checks` es `[]`, `document` es `null` y
        `expires_at` es `null`. En un estado terminal, `upload` es `null` y `expires_at` informa
        cuándo se elimina el resultado.
      required:
        - id
        - object
        - sandbox
        - document_type
        - face_match
        - status
        - declined_reason
        - checks
        - document
        - autorizacion
        - upload
        - webhook_url
        - return_url
        - hosted_url
        - created_at
        - updated_at
        - completed_at
        - expires_at
      additionalProperties: false
      properties:
        id:
          type: string
          pattern: "^val_[0-9a-f]{32}$"
        object:
          type: string
          const: validation
        sandbox:
          type: boolean
        document_type:
          $ref: "#/components/schemas/DocumentType"
        face_match:
          type: boolean
        status:
          $ref: "#/components/schemas/ValidationStatus"
        declined_reason:
          $ref: "#/components/schemas/DeclinedReason"
        checks:
          type: array
          maxItems: 5
          items:
            $ref: "#/components/schemas/Check"
        document:
          oneOf:
            - $ref: "#/components/schemas/Document"
            - type: "null"
        autorizacion:
          $ref: "#/components/schemas/AutorizacionRegistrada"
        upload:
          oneOf:
            - $ref: "#/components/schemas/Upload"
            - type: "null"
        webhook_url:
          type: [string, "null"]
          format: uri
          maxLength: 2048
        return_url:
          type: [string, "null"]
          maxLength: 2048
          description: Retorno de la sesión alojada pedido al crear, o `null`.
        hosted_url:
          type: [string, "null"]
          format: uri
          pattern: "/v/[A-Za-z0-9_-]{43,}$"
          description: |
            Página alojada `<URL_PUBLICA>/v/<token>` para leer el documento sin integrar el
            componente. Su token es distinto del de `upload.url`, identifica la validación y vence con
            `upload.expires_at`. `null` fuera de `pending`.
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: [string, "null"]
          format: date-time
        expires_at:
          type: [string, "null"]
          format: date-time
    ImagenesValidacion:
      type: object
      description: Partes de la subida; cada una JPEG o PNG de 1 a 8 388 608 bytes.
      required: [front, back]
      additionalProperties: false
      properties:
        front:
          type: string
          format: binary
          description: Anverso del documento.
        back:
          type: string
          format: binary
          description: Reverso del documento.
        selfie:
          type: string
          format: binary
          description: Solo si la validación pidió `face_match`; prohibida en otro caso.
...